+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ------------------------------------------------------
-Create: Mobile Packages - v0.6.1 - 1.21.1 - unreleased
+Create: Mobile Packages - v0.7.0 - 1.21.1 - unreleased
+------------------------------------------------------
+
+### Additions
+
+- Added TrashSlots to the Portable Stock Ticker allowing players to send items (#299)
+    - Set an address
+    - Put items in the TrashSlots
+    - A Robo Bee will come and pick up the items and send them to the address
+
+------------------------------------------------------
+Create: Mobile Packages - v0.6.1 - 1.21.1 - 20.02.2026
 ------------------------------------------------------
 
 ### Bug Fixes
@@ -8,7 +19,7 @@ Create: Mobile Packages - v0.6.1 - 1.21.1 - unreleased
 - Add Bee Port clearing recipe
 
 ### Changes
-- Update Translations (#285)
+- Update Translations (#285, #297)
 
 ------------------------------------------------------
 Create: Mobile Packages - v0.6.0 - 1.21.1 - 15.02.2026